requirements.txt
Flask==2.3.3
openai-whisper==20231117
werkzeug==2.3.7
gunicorn